finished adding open source engines

+++ erp_app/lib/erp_app/extensions/active_record/acts_as_app_container.rb
@@ -59,8 +59,7 @@ module ErpApp
 				  end
 
 				  def destroy_app_container
-					if self.app_container && !self.app_container.frozen?
-					  self.app_container.destroy
+					  self.app_container.destroy if (self.app_container && !self.app_container.frozen?)
 					end
 				  end
